test(cli/governance:propose): refactors multisig creation into a function
